feat:Edit and delete comment, register, User/follower/following, perm…

+++ src/blog/views/views_v1.py
@@ -3,18 +3,49 @@ from rest_framework.response import Response
 from rest_framework import status
 from ..serializers import (ArticleReadSerializer, ArticleDetailSerializer, 
                           ArticlesCommentSerializer, ArticleWriteSerializer, ArticleUpdateSerializer, 
-                          AllCommentSerializer, BookMarkSerializer, CategoryReadSerializer, CategoryWriteSerializer, ContactUsSerializer)
+                          AllCommentSerializer, BookMarkWriteSerializer, CategoryReadSerializer,BookMarkReadSerializer, 
+                          CategoryWriteSerializer, ContactUsSerializer, UpdateCommentSerializer)
 from ..models import ArticleModel, CommentModel, BookMarkModel, CategoryModel
 from django.shortcuts import get_object_or_404
 from django.db.models import Count
-from ..pagination import ArticlePagination
+from ..pagination import MyPagination
 from account.models import CustomUser
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from rest_framework.permissions import IsAuthenticatedOrReadOnly, IsAuthenticated
 from django.db.models import Q, Prefetch
 import mimetypes
 from django.core.exceptions import ValidationError
+from drf_yasg.utils import swagger_auto_schema
+from rest_framework.parsers import MultiPartParser, FormParser
+from ..custom_permissions import IsArticleAuthorOrReadOnly, IsCommentAuthorOrReadOnly
+from drf_yasg import openapi
 
 class ArticleBaseView(APIView):
+    parser_classes = [MultiPartParser, FormParser]
+    permission_classes=[IsAuthenticatedOrReadOnly]
+
+    @swagger_auto_schema(
+        manual_parameters=[
+            openapi.Parameter(
+                "follower_fullname",
+                openapi.IN_QUERY,
+                description="Filter followers by full name (case-insensitive match).",
+                type=openapi.TYPE_STRING,
+            ),
+            openapi.Parameter(
+                "page",
+                openapi.IN_QUERY,
+                description="Specify the page number for pagination.",
+                type=openapi.TYPE_INTEGER,
+            ),
+        ],
+        responses={
+            200: openapi.Response(
+                description="A list of followers.",
+                schema=ArticleReadSerializer(many=True),
+            ),
+            400: "Bad Request",
+        }
+    )
     def get(self, request):
         query_params = dict(request.query_params)
         search=query_params.get('search')  # get search from the request
@@ -28,16 +59,19 @@ class ArticleBaseView(APIView):
             categories=[int(i) for i in categories]
             articles = articles.filter(categories__id__in=categories)
 
-        paginator = ArticlePagination()
+        paginator = MyPagination()
         paginated_articles = paginator.paginate_queryset(articles, request)
         serializers=ArticleReadSerializer(paginated_articles, many=True)
         return paginator.get_paginated_response(serializers.data)
     
+
+    @swagger_auto_schema(
+    request_body=ArticleWriteSerializer,
+    responses={201: ArticleReadSerializer()})
     def post(self, request):
-        request_user = CustomUser.objects.get(id=1)
         serializer = ArticleWriteSerializer(data=request.data)
         if serializer.is_valid(raise_exception=True):
-            serializer.save(author=request_user)
+            serializer.save(author=request.user)
             return Response(
                     data=serializer.data,
                     status=status.HTTP_201_CREATED)
@@ -45,10 +79,11 @@ class ArticleBaseView(APIView):
         
    
 class ArticleDetailView(APIView):
-
+    permission_classes = [IsAuthenticatedOrReadOnly, IsArticleAuthorOrReadOnly] 
     def get_object(self, slug):
         return get_object_or_404(ArticleModel, slug=slug)
 
+    @swagger_auto_schema(responses={200: ArticleDetailSerializer()})
     def get(self, request, slug):
         article = self.get_object(slug)
         serializer = ArticleDetailSerializer(article)
@@ -58,16 +93,21 @@ class ArticleDetailView(APIView):
         )
 
     def delete(self, request, slug):
-        # permission_classes = [IsAuthenticatedOrReadOnly]
         article = self.get_object(slug)
+        self.check_object_permissions(request, article) # explicitly called
         article.delete()
         return Response(
             data={"success": True},
             status=status.HTTP_204_NO_CONTENT
         )
+    @swagger_auto_schema(
+        request_body=ArticleUpdateSerializer,
+        responses={200: ArticleReadSerializer()})
     
     def patch(self, request, slug):
         article = self.get_object(slug)
+        print(f"Request user: {request.user}, Article author: {article.author}") 
+        self.check_object_permissions(request, article) # explicitly called
         serializer = ArticleUpdateSerializer(data=request.data, instance=article, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
@@ -75,9 +115,14 @@ class ArticleDetailView(APIView):
                     data=serializer.data,
                     status=status.HTTP_201_CREATED
                     )
-
+        
+    @swagger_auto_schema(
+        request_body=ArticleUpdateSerializer,
+        responses={200: ArticleReadSerializer()})
+    
     def put(self, request, slug):
         article = self.get_object(slug)
+        self.check_object_permissions(request, article) # explicitly called
         serializer = ArticleUpdateSerializer(data=request.data, instance=article)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
@@ -86,8 +131,11 @@ class ArticleDetailView(APIView):
                     status=status.HTTP_201_CREATED
                     )
 
-    
+@swagger_auto_schema(responses={200: ArticlesCommentSerializer(many=True)})    
 class MainCommentView(APIView):
+    permission_classes=[IsAuthenticatedOrReadOnly, IsCommentAuthorOrReadOnly]
+
+    @swagger_auto_schema(responses={200: ArticlesCommentSerializer(many=True)})
     def get(self, request, slug):
         # Prefetch only the main comments (no parent)
         main_comments = Prefetch('comments', queryset=CommentModel.objects.filter(parent__isnull=True))
@@ -105,8 +153,48 @@ class MainCommentView(APIView):
             data=serializer.data,
             status=status.HTTP_200_OK
         )
-
     
+    def get_article_comment(self, slug, comment_id):
+        article = get_object_or_404(ArticleModel, slug=slug)
+        comment = article.comments.get(id=comment_id)
+        return article, comment
+
+
+    @swagger_auto_schema(
+        responses={
+            204: 'Comment was deleted successfully',
+        }
+    )
+    def delete(self, request, slug, comment_id):
+        # ignore article (return value of the helper method) by using _
+        _, comment = self.get_article_comment(slug, comment_id)
+        self.check_object_permissions(request, comment) # explicitly called
+        # Delete the comment
+        comment.delete()
+        return Response(
+                data={"success": True},
+                status=status.HTTP_204_NO_CONTENT)
+    
+
+    @swagger_auto_schema(
+        request_body=UpdateCommentSerializer,
+        responses={
+            200: UpdateCommentSerializer,
+            400: 'Bad request, invalid data.',
+        }
+    )
+    def put(self, request, slug, comment_id):
+        _, comment = self.get_article_comment(slug, comment_id)
+        self.check_object_permissions(request, comment) # explicitly called
+        serializer=UpdateCommentSerializer(comment, data=request.data)
+        if serializer.is_valid(raise_exception=True):
+            serializer.save()
+            return Response(
+                data=serializer.data,
+                status=status.HTTP_200_OK)
+       
+
+@swagger_auto_schema(responses={200: AllCommentSerializer(many=True)})    
 class AllCommentsView(APIView):
     def get(self, request, id):
         # Retrieve the comment by its ID
@@ -127,19 +215,17 @@ class AllCommentsView(APIView):
 
 class BookMarkView(APIView):
     def post(self, request, article_id):
-        request_user = CustomUser.objects.get(id=3)
-
         # Check if the user has already bookmarked this article
-        if BookMarkModel.objects.filter(user=request_user, article__id=article_id).exists():
+        if BookMarkModel.objects.filter(user=request.user, article__id=article_id).exists():
             return Response({"detail": "You have already bookmarked this article."}, status=status.HTTP_400_BAD_REQUEST)
         
-        serializer = BookMarkSerializer(data={"article": article_id})
+        serializer = BookMarkWriteSerializer(data={"article": article_id})
         if serializer.is_valid(raise_exception=True):
-            serializer.save(user=request_user)
+            serializer.save(user=request.user)
             return Response(
                     data=serializer.data,
                     status=status.HTTP_201_CREATED)
-    
+        
     
     def delete(self, request, article_id):
         article = get_object_or_404(BookMarkModel, article_id=article_id)
@@ -148,26 +234,60 @@ class BookMarkView(APIView):
             data={"success": True},
             status=status.HTTP_204_NO_CONTENT
         )
+
+    def get(self, request):
+        article_slug = request.query_params.get('article_slug')
+        user= request.user
+        bookmarks=BookMarkModel.objects.filter(user=user)
+
+        if article_slug:
+            bookmarks = bookmarks.filter(article__slug__icontains=article_slug)
+
+        paginator = MyPagination()
+        paginated_bookmarks=paginator.paginate_queryset(bookmarks, request) 
+        serializer=BookMarkReadSerializer(paginated_bookmarks, many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
     
-class CategoryView (APIView):
-    def post(self, request):
-        serializer = CategoryWriteSerializer(data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(
-                    data=serializer.data,
-                    status=status.HTTP_201_CREATED
-                    )
+# class CategoryView (APIView):
+#     def post(self, request):
+#         serializer = CategoryWriteSerializer(data=request.data)
+#         if serializer.is_valid(raise_exception=True):
+#             serializer.save()
+#             return Response(
+#                     data=serializer.data,
+#                     status=status.HTTP_201_CREATED
+#                     )
         
-    def delete(self, request, id):
-        category = get_object_or_404(CategoryModel, id=id)
-        category.delete()
-        return Response(
-            data={"success": True},
-            status=status.HTTP_204_NO_CONTENT
-        )
+#     def delete(self, request, id):
+#         category = get_object_or_404(CategoryModel, id=id)
+#         category.delete()
+#         return Response(
+#             data={"success": True},
+#             status=status.HTTP_204_NO_CONTENT
+#         )
+    
+#     def get(self, request):
+#         categories=CategoryModel.objects.all()
+#         category_slug = request.query_params.get('category_slug')
+        
+#         if category_slug:
+#             categories = categories.filter(slug__icontains=category_slug)
+
+#         paginator = MyPagination()
+#         paginated_categories=paginator.paginate_queryset(categories, request) 
+#         serializer=CategoryReadSerializer(paginated_categories, many=True)
+#         return Response(serializer.data, status=status.HTTP_200_OK)
+
 
 class ContactUsView(APIView):
+    @swagger_auto_schema(
+    request_body=ContactUsSerializer,
+    responses={
+        201: ContactUsSerializer,  # This will show the serializer data as response
+        400: 'Bad request, invalid input data.'
+    }
+)
     def post (self, request):
         serializer = ContactUsSerializer(data=request.data)
         if serializer.is_valid(raise_exception=True):
@@ -179,8 +299,7 @@ class ContactUsView(APIView):
 
 class MyArticlesView(APIView):
     def get(self, request):
-        request_user = CustomUser.objects.get(id=1)
-        my_articles= ArticleModel.objects.filter(author=request_user)
+        my_articles= ArticleModel.objects.filter(author=request.user)
         serializer = ArticleReadSerializer(my_articles, many=True)
         return Response(
             data=serializer.data,
